# Use new static HashData methods for .NET 6+ targets