--- a/osu.Framework/Extensions/ExtensionMethods.cs
+++ b/osu.Framework/Extensions/ExtensionMethods.cs
@@ -269,9 +269,10 @@ namespace osu.Framework.Extensions
         {
 #if NET6_0_OR_GREATER
             return SHA256.HashData(Encoding.UTF8.GetBytes(str)).toLowercaseHex();
-#endif
+#else
             using (var alg = SHA256.Create())
                 return alg.ComputeHash(Encoding.UTF8.GetBytes(str)).toLowercaseHex();
+#endif
         }
 
         public static string ComputeMD5Hash(this Stream stream)
@@ -290,9 +291,10 @@ namespace osu.Framework.Extensions
         {
 #if NET6_0_OR_GREATER
             return MD5.HashData(Encoding.UTF8.GetBytes(input)).toLowercaseHex();
-#endif
+#else
             using (var md5 = MD5.Create())
                 return md5.ComputeHash(Encoding.UTF8.GetBytes(input)).toLowercaseHex();
+#endif
         }
 
         public static DisplayIndex GetIndex(this DisplayDevice display)