Use the new static one-shot APIs introduced in .NET 5 to improve performance and allocations.

Benchmark numbers on my Windows machine:
Before:
|    Method |       Mean |    Error |   StdDev |  Gen 0 | Allocated |
|---------- |-----------:|---------:|---------:|-------:|----------:|
| StringMD5 |   905.6 ns | 13.56 ns | 12.68 ns | 0.0839 |     352 B |
| StringSHA | 1,060.0 ns |  2.35 ns |  2.19 ns | 0.1068 |     448 B |

After:
|    Method |     Mean |   Error |  StdDev |  Gen 0 | Allocated |
|---------- |---------:|--------:|--------:|-------:|----------:|
| StringMD5 | 648.2 ns | 9.07 ns | 8.04 ns | 0.0439 |     184 B |
| StringSHA | 820.9 ns | 4.85 ns | 4.30 ns | 0.0629 |     264 B |